+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Update dependencies
+- Prefer eaton configuration via options over `-c` option
 
 ## 1.5.2 - 2022-10-06
 - Fix creating an Api